[bugfix] Add non-linux default implementation of live query code that will cleanly fail
